// src/types/trend.ts
export type Trend = {
	imp: number;
	click: number;
	cost: number;
	conv: number;
	convValue: number;
	ctr: number;
	cvr: number;
	cpc: number;
	cpa: number;
	roas: number;
	date: string;
};

/**
 * imp : 노출수
 * click : 클릭수
 * cost : 비용
 * conv : 전환수
 * convValue : 전환가치
 * ctr : 클릭률
 * cvr : 전환률(구매 전환률)
 * cpc : 클릭당 비용
 * cpa : 참여당 비용
 * roas : 광고수익률
 * date: 해당 일자
 */